[드랍과 인벤토리] 저장할 아이템의 모델(다이나믹 데이터)인 ItemModel에 uniqueId를 저장할 수 있도록 long형으로 변경
- long -> 서버/인벤토리에 절대 겹치지 않게 식별하려고 안전한 타입으로 바꿈

+++ Assets/Scripts/ViewModel/PlayerModel.cs
@@ -23,7 +23,7 @@ public class PlayerModel
 [Serializable]
 public class ItemViewModel
 {
-    public int ItemUniqueId;    // 아이템 고유 ID
+    public long ItemUniqueId;    // 아이템 고유 ID
     public string ItemDataId;   // 어떤 종류의 아이템인지 알려주는 ID
     public int ItemStackCount;  // 아이템 개수
 }
